Added formik props to Select component, moved component into Forms folder
