+++ resources/assets/js/components/ApplicationReview/ApplicationReviewWithNav.tsx
@@ -8,7 +8,7 @@ import {
 import className from "classnames";
 import Swal from "sweetalert2";
 import * as routes from "../../helpers/routes";
-import Select, { SelectOption } from "../Select";
+import Select, { SelectOption } from "../Forms/Select";
 import { Application } from "../../models/types";
 import { ReviewStatusId } from "../../models/lookupConstants";
