+++ resources/assets/js/components/ApplicationReview/ReviewApplications.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import moment from "moment";
 import { FormattedMessage } from "react-intl";
 import { Application } from "../../models/types";
-import { SelectOption } from "../Select";
+import { SelectOption } from "../Forms/Select";
 import { applicationCategory } from "./helpers";
 import ReviewCategory from "./ReviewCategory";
 
@@ -120,7 +120,10 @@ const ReviewApplications: React.StatelessComponent<ReviewApplicationsProps> = ({
             description="Welcome header on app main page"
             values={{
               // TODO: Think more carefully about how to handle null fields
-              dayCount: moment().diff(closeDateTime ? moment(closeDateTime) : moment(), "days"),
+              dayCount: moment().diff(
+                closeDateTime ? moment(closeDateTime) : moment(),
+                "days",
+              ),
             }}
           />
         </div>
