+++ resources/assets/js/components/Forms/Select.tsx
@@ -15,6 +15,11 @@ export interface SelectProps<T extends string | number> {
   options: SelectOption<T>[];
   onChange: (event: React.ChangeEvent<HTMLSelectElement>) => void;
   errorText?: string;
+  grid?: string;
+
+  // formik
+  field?: any;
+  form?: any;
 }
 
 export default function Select<T extends string | number>(
@@ -30,19 +35,28 @@ export default function Select<T extends string | number>(
     options,
     onChange,
     errorText,
+    grid,
+    field,
+    form: { errors },
   } = props;
 
+  const { name } = field;
+
   return (
-    <div data-c-input="select">
+    <div
+      data-c-input="select"
+      data-c-grid-item={grid}
+      data-c-required={required}
+    >
       <label htmlFor={htmlId}>{label}</label>
       {required && <span>Required</span>}
       <div>
         <i className="fa fa-caret-down" />
         <select
           id={htmlId}
-          name={formName}
+          name={name || formName}
           value={selected || ""}
-          onChange={e => onChange(e)}
+          onChange={field.onChange || (e => onChange(e))}
         >
           {nullSelection && (
             <option value="" disabled>
@@ -56,7 +70,7 @@ export default function Select<T extends string | number>(
           ))}
         </select>
       </div>
-      <span>{errorText || "This input has an error."}</span>
+      <span>{errors[name] || errorText || "This input has an error."}</span>
     </div>
   );
 }
